feat: rest api versioning
add /v1 prefix to endpoints. Route definitions are extracted into
rest-api/src/v1/routes.rs and nested via Router::nest("/v1", …), keeping
handlers and middleware shared.
