+++ rest-api/src/v1.rs
@@ -1,0 +1,1 @@
+pub mod routes;
